+++ NEWS.md
@@ -10,6 +10,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Display dialog with condition details 
+  [#16](https://github.com/orontee/argos/issues/16)
+
 - Specify language when calling OpenWeather API
   [#15](https://github.com/orontee/argos/issues/15)